+++ src/model/index.js
@@ -3,4 +3,4 @@ export { default as LottoMachine } from "./LottoMachine.js";
 export { default as WinningLotto } from "./WinningLotto.js";
 export { default as LottoNumber } from "./LottoNumber.js";
 export { default as LottoMatcher } from "./LottoMatcher.js";
-export { default as LottoResult } from "./LottoResult.js";
+export { default as LottoReturnCalculator } from "./LottoReturnCalculator.js";
